+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 -----------------------
 
 - added try / catch to :meth:`~import_export.results.RowResult.add_instance_info` to handle unserializable instances (#1767)
+- fix: YAML export does not work with SafeString (#1762)
 
 4.0.0-rc.0 (2024-02-14)
 -----------------------